Time: 252 ms (17.16%), Space: 83.9 MB (25.19%) - LeetHub

+++ 525-contiguous-array/525-contiguous-array.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int findMaxLength(vector<int>& nums) {
-        unordered_map<float, int> m;
+        unordered_map<int, int> m;
         m[0] = -1;
         int sum = 0, ans = 0;
         for(int i = 0;i<nums.size();i++)
